Web: review fixes — clamp toInt, Shift+Enter direction, settings comment
- web_find.dart: int.clamp returns num; .toInt() for the parse index.
- browser_find_bar.dart: the TextField consumes Enter before the
  wrapping Focus sees it, so Shift+Enter never reached the dead branch —
  read the direction in onSubmitted via HardwareKeyboard instead.
- settings_dialog.dart: stale 'two settings' comment now covers three.

+++ lib/core/web/web_find.dart
@@ -120,7 +120,9 @@ class WebFindResult {
       final Object? t = raw['total'];
       final Object? i = raw['index'];
       if (t is int && i is int && t >= 0) {
-        return WebFindResult(total: t, index: i.clamp(0, t));
+        // `clamp` answers `num` even for ints — the `.toInt()` is load-
+        // bearing, not decoration.
+        return WebFindResult(total: t, index: i.clamp(0, t).toInt());
       }
     }
     return none;
